pyproject.toml
[project]
name = "athena"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.46.0",
    "pypdf>=6.1.3",
    "python-dotenv>=1.2.1",
]
